% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_addons.R
\name{excd}
\alias{excd}
\title{Exceedance Probabilities}
\usage{
excd(gaugedata, discharge = "vattenforing_m3_s", constant = 0.3)
}
\arguments{
\item{gaugedata}{data.frame with discharge column}

\item{discharge}{A column with discharge data}

\item{constant}{numeric. Gringorten plotting position formula}
}
\value{
exceedence table
}
\description{
This function calculates exceedance probabilities using Gringorten plotting position formula.
}
