% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_addons.R
\name{ff_vals}
\alias{ff_vals}
\title{Get FF Values}
\usage{
ff_vals(data, value_name, wy_month = 10)
}
\arguments{
\item{data}{A data.frame with date and flow columns.}

\item{value_name}{One unquoted expression for flow column name, e.g. vettenforing_m3_s.}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\value{
A \code{tibble} with Q1, Q2, and Q5 flood frequency values
}
\description{
Get FF Values
}
