% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_addons.R
\name{floowy}
\alias{floowy}
\title{Get Peak Flows}
\usage{
floowy(data, value_name, wy_month = 10)
}
\arguments{
\item{data}{A data.frame with date and flow columns.}

\item{value_name}{One unquoted expression for flow column name, e.g. vettenforing_m3_s.}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\value{
A data.frame with peak flows labelled `peak_flow`.
}
\description{
Get Peak Flows
}
