% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_addons.R
\name{get_rbi}
\alias{get_rbi}
\title{Richards-Baker Flashiness index (RBI)}
\usage{
get_rbi(data, flow_var, ...)
}
\arguments{
\item{data}{A daily value data.frame.}

\item{flow_var}{A column name for flow, e.g. Flow.}

\item{...}{values to pass to `dplyr::group_by()` to perform RBI calculation on. (optional)}
}
\value{
A rbi column within data.frame.
}
\description{
Adds RBI of observation (flow).
}
\note{
The user must specify the `flow_var` column but it's optional whether to use `dplyr::group_by()` or not.
}
\examples{
stream_flow <- data.frame(flow = c(seq(30, 60), seq(60, 30, length.out = 60)))
stream_flow \%>\% get_rbi(flow)
}
