% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{getRanges}
\alias{getRanges}
\title{Get Ranges}
\usage{
getRanges(expand = 1.03, ranges = rgl::par3d("bbox"))
}
\arguments{
\item{expand}{numeric}

\item{ranges}{rgl par3d function.}
}
\value{
A range of x, y, and z values.
}
\description{
This function is taken from the rgl package due to it being non-exported.
}
