% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{latitude_rastergraph}
\alias{latitude_rastergraph}
\title{Latitude Rastergraph}
\usage{
latitude_rastergraph(data)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Latitude Rastergraph
}
\note{
Very specific right now with what goes in because we need to join the station metadata latitudes.
}
