% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_addons.R
\name{prep_flow}
\alias{prep_flow}
\title{Prepping Flow Data}
\usage{
prep_flow(data, value_name = vattenforing_m3_s, wy_month = 10)
}
\arguments{
\item{data}{A data.frame with date and flow columns}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\description{
Prepping Flow Data
}
\note{
In data.frame `date` column must be named `date`.
}
