% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{riverwave_2d}
\alias{riverwave_2d}
\title{2D plot of riverwave with breakpoints}
\usage{
riverwave_2d(flowdata, doy, discharge, year, breakpoints)
}
\arguments{
\item{flowdata}{dataframe with discharge column, day of year column, year column}

\item{doy}{day of year column pointer}

\item{discharge}{discharge column pointer}

\item{year}{year column pointer}

\item{breakpoints}{a vector with 3 positions for gradient, 1st pos. is usually 1}
}
\description{
This plotting function returns a ggplot of daily discharge colored by riverwave breakpoints across time
}
