% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{riverwave_percentiles_plot}
\alias{riverwave_percentiles_plot}
\title{Get DOY Summary Stats Graph}
\usage{
riverwave_percentiles_plot(data, value_name, wy_month = 10)
}
\arguments{
\item{data}{A data.frame with date and flow columns.}

\item{value_name}{One unquoted expression for flow column name, e.g. vettenforing_m3_s.}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\value{
A \code{ggplot} with percentiles, proportion to wy max and Q1, Q2, Q5 flood frequency values.
}
\description{
Get DOY Summary Stats Graph
}
