% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{rw_2d}
\alias{rw_2d}
\title{2D plot of riverwave with breakpoints}
\usage{
rw_2d(data, discharge, breakpoints, wy_month = 10)
}
\arguments{
\item{data}{dataframe with discharge and `date` column.}

\item{discharge}{discharge column pointer}

\item{breakpoints}{a vector with 3 positions for gradient, 1st pos. is usually 1}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\description{
This plotting function returns a ggplot of daily discharge colored by riverwave breakpoints across time
}
\note{
In data.frame or \code{tibble} `date` column must be named `date` and is of origin "%Y-%m-%d".
}
