% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{rw_hydrograph}
\alias{rw_hydrograph}
\title{Raster Hydrograph}
\usage{
rw_hydrograph(data, value_name)
}
\arguments{
\item{data}{A data.frame with date and flow columns.}

\item{value_name}{The variable to include as the y-axis.}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\value{
A ggplot.
}
\description{
Raster Hydrograph
}
\note{
The `date` column must be named `date` and is of origin "%Y-%m-%d".
}
\examples{
\dontrun{
#provide a station that you want to look at visually (we need a date and flow column)
site_10 <- smhiDataClean[['10']]

raster_hydrograph(data = site_10, value_name = vattenforing_m3_s)

}

}
