% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{rw_latitude_rastergraph}
\alias{rw_latitude_rastergraph}
\title{Latitude Rastergraph}
\usage{
rw_latitude_rastergraph(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A plot.
}
\description{
Latitude Rastergraph
}
\note{
Very specific right now with what goes in because we need to join the station metadata latitudes.
}
