% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{rw_percentiles_plot}
\alias{rw_percentiles_plot}
\title{Get DOY Summary Stats Graph}
\usage{
rw_percentiles_plot(data, value_name, q1, q2, q5, wy_month = 10)
}
\arguments{
\item{data}{A data.frame with date and flow columns.}

\item{value_name}{One unquoted expression for flow column name, e.g. vettenforing_m3_s.}

\item{q1}{Numeric. Peakflow for 1 year recurrence interval.}

\item{q2}{Numeric. Peakflow for 2 year recurrence interval.}

\item{q5}{Numeric. Peakflow for 5 year recurrence interval.}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\value{
A \code{ggplot} with percentiles, proportion to wy max and Q1, Q2, Q5 flood frequency values.
}
\description{
Get DOY Summary Stats Graph
}
