% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{rw_plot3D}
\alias{rw_plot3D}
\title{Plot 3D for RW}
\usage{
rw_plot3D(
  x,
  maxpixels = 1e+05,
  zfac = 1,
  drape = NULL,
  col = terrain.colors,
  at = 100,
  rev = FALSE,
  useLegend = TRUE,
  adjust = TRUE,
  ...
)
}
\arguments{
\item{x}{vectors of points to be plotted. Any reasonable way of defining the coordinates is acceptable. See the function xyz.coords for details.}

\item{maxpixels}{A numeric.}

\item{zfac}{A numeric.}

\item{drape}{A logical}

\item{col}{A vector of colors}

\item{at}{A numeric}

\item{rev}{A logical}

\item{useLegend}{A logical}

\item{adjust}{A logical}

\item{...}{Arguments to pass to \link[rgl]{surface3d}.}
}
\value{
A 3d riverwave plot.
}
\description{
Plot 3D for RW
}
