% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_vatten.R
\name{smhi_vatten}
\alias{smhi_vatten}
\title{Get SMHI Vatten}
\usage{
smhi_vatten(stn_no)
}
\arguments{
\item{stn_no}{Numeric. Station Number.}
}
\value{
A data.frame
}
\description{
Get SMHI Vatten
}
\examples{
{

ellinge_vatten <- smhi_vatten(1132)

}

}
