% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_vatten.R
\name{smhi_vatten_current}
\alias{smhi_vatten_current}
\title{Title}
\usage{
smhi_vatten_current(inv)
}
\arguments{
\item{inv}{dataframe of smhi inventory}
}
\value{
station numbers currently monitored
}
\description{
Title
}
\examples{

}
