% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_vatten.R
\name{smhi_vatten_data}
\alias{smhi_vatten_data}
\title{Get SMHI Vatten}
\usage{
smhi_vatten_data(stn_no)
}
\arguments{
\item{stn_no}{Numeric. Station Number.}
}
\value{
A data.frame
}
\description{
Get SMHI Vatten
}
\examples{

}
