% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_vatten.R
\name{smhi_vatten_map}
\alias{smhi_vatten_map}
\title{SMHI Vatten Map}
\usage{
smhi_vatten_map(smhi_vatten_inv)
}
\arguments{
\item{smhi_vatten_inv}{dataframe of gauge inventory from smhi_vatten_inv}
}
\value{
A leaflet map.
}
\description{
SMHI Vatten Map
}
