% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_vatten.R
\name{smhi_vatten_natural}
\alias{smhi_vatten_natural}
\title{SMHI Natural Model}
\usage{
smhi_vatten_natural(stn_no)
}
\arguments{
\item{stn_no}{Numeric. Station Number.}
}
\value{
dataframe of natural modeled data from 2010-10-01 to 2023-09-30.
}
\description{
SMHI Natural Model
}
