% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_vatten.R
\name{smhi_vatten_natural_pt}
\alias{smhi_vatten_natural_pt}
\title{SMHI Natural Model (spatial)}
\usage{
smhi_vatten_natural_pt(point)
}
\arguments{
\item{point}{A sf object.}
}
\value{
dataframe of natural modeled data from 2010-10-01 to 2023-09-30.
}
\description{
SMHI Natural Model (spatial)
}
