% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_addons.R
\name{summary_stats_doy}
\alias{summary_stats_doy}
\title{Get DOY Summary Stats}
\usage{
summary_stats_doy(data, value_name, wy_month = 10)
}
\arguments{
\item{data}{A data.frame with date and flow columns.}

\item{value_name}{One unquoted expression for flow column name, e.g. vettenforing_m3_s.}

\item{wy_month}{A numeric for what month to use as start of water year, 10 (default).}
}
\value{
A \code{tibble} with percentiles of `wy_doy`.
}
\description{
Get DOY Summary Stats
}
