% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ternary}
\alias{ternary}
\title{ternary}
\usage{
ternary(data, coefv, rbi, bfi, lat, cat_size)
}
\arguments{
\item{data}{}
}
\value{
a ternary plot with coefv, rbi, and bfi axes
}
\description{
ternary
}
\examples{
ternary(test)


}
