% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverwave_plot.R
\name{rw_plot3D}
\alias{rw_plot3D}
\title{Plot 3D for RW}
\usage{
rw_plot3D(
  x,
  maxpixels = 1e+05,
  zfac = 1,
  drape = NULL,
  col = terrain.colors,
  at = 100,
  rev = FALSE,
  useLegend = TRUE,
  adjust = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Plot 3D for RW
}
