% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smhi_vatten.R
\name{smhi_vatten}
\alias{smhi_vatten}
\title{Get all data from SMHI Vatten}
\usage{
smhi_vatten(stn_nos)
}
\arguments{
\item{stn_nos}{vector of Numeric. Station Number.}
}
\value{
A data.frame
}
\description{
Get all data from SMHI Vatten
}
\examples{
{

ellinge_vatten <- smhi_vatten(1132)

}

}
