% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary.R
\name{ternary}
\alias{ternary}
\title{ternary}
\usage{
ternary(data, coefv, rbi, bfi, lat, cat_size)
}
\arguments{
\item{data}{dataframe of measurements}

\item{coefv}{coefficient of variation}

\item{rbi}{richard baker index}

\item{lat}{latitude}

\item{cat_size}{catchment size}
}
\value{
a ternary plot with coefv, rbi, and bfi axes
}
\description{
ternary
}
